Abstracted out socket.io test into a component to make ready for building out app.
